add permissions for derived modules

+++ apps/warehouse/src/meqs-supplier-attachment/meqs-supplier-attachment.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Logger, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { CreateMeqsSupplierAttachmentInput } from './dto/create-meqs-supplier-attachment.input';
 import { UpdateMeqsSupplierAttachmentInput } from './dto/update-meqs-supplier-attachment.input';
 import { PrismaService } from '../__prisma__/prisma.service';
@@ -6,6 +6,7 @@ import { MEQSSupplierAttachment, Prisma } from 'apps/warehouse/prisma/generated/
 import { WarehouseRemoveResponse } from '../__common__/classes';
 import { AuthUser } from '../__common__/auth-user.entity';
 import axios from 'axios';
+import { isAdmin } from '../__common__/helpers';
 
 @Injectable()
 export class MeqsSupplierAttachmentService {
@@ -21,6 +22,9 @@ export class MeqsSupplierAttachmentService {
 
     async create(input: CreateMeqsSupplierAttachmentInput): Promise<MEQSSupplierAttachment> {
 
+        if (!this.canAccess(input.meqs_supplier_id)) {
+            throw new ForbiddenException('Only Admin and Owner can create meqs supplier attachment!')
+        }
 
         const data: Prisma.MEQSSupplierAttachmentCreateInput = {
             meqs_supplier: { connect: { id: input.meqs_supplier_id } },
@@ -64,6 +68,10 @@ export class MeqsSupplierAttachmentService {
 
         const existingItem = await this.findOne(id)
 
+        if (!this.canAccess(existingItem.meqs_supplier_id)) {
+            throw new ForbiddenException('Only Admin and Owner can update meqs supplier!')
+        }
+
         const data: Prisma.MEQSSupplierAttachmentUpdateInput = {
             src: input.src ?? existingItem.src,
             filename: input.filename ?? existingItem.filename,
@@ -88,6 +96,10 @@ export class MeqsSupplierAttachmentService {
 
         const existingItem = await this.findOne(id)
 
+        if (!this.canAccess(existingItem.meqs_supplier_id)) {
+            throw new ForbiddenException('Only Admin and Owner can remove meqs supplier!')
+        }
+
         await this.prisma.mEQSSupplierAttachment.delete({
             where: { id },
         })
@@ -110,6 +122,31 @@ export class MeqsSupplierAttachmentService {
         console.log('url', url)
 
         return axios.delete(url, { data: filePaths });
+    }
+
+    private async canAccess(meqs_supplier_id: string): Promise<boolean> {
+
+        if (isAdmin(this.authUser)) return true
+
+        const meqsSupplier = await this.prisma.mEQSSupplier.findUnique({
+            where: {
+                id: meqs_supplier_id
+            },
+            include: {
+                meqs: true
+            }
+        })
+
+        if (!meqsSupplier) {
+            throw new NotFoundException('meqsSupplier not found with id of ' + meqs_supplier_id)
+        }
+
+        const isOwner = meqsSupplier.meqs.created_by === this.authUser.user.username
+
+        if (isOwner) return true
+
+        return false
+
     }
 
 }
